+++ tests/test_formatter/test_formatter_output.py
@@ -158,6 +158,8 @@ def test_ipynb(snapshot):
             'flake8',
             filename,
             '--disable-noqa',
+            '--select',
+            'WPS',
             '--isolated',
             '--format',
             'wemake',